fix unit tests and start refactoring
