+++ mcda/mcda_functions/aggregation.py
